+++ springDream/springDream/APP/ShopCart(购物车)/Models/HHCartModel.m
@@ -68,27 +68,27 @@
 }
 - (void)mj_keyValuesDidFinishConvertingToObject{
     
-    NSString *protocolStr = [NSString stringWithFormat:@"¥%.2f",self.showMoney.floatValue];
-    NSString *contentStr = [NSString stringWithFormat:@"共%ld件商品,合计:¥%.2f",self.products.count,self.showMoney.floatValue];
+    NSString *protocolStr = [NSString stringWithFormat:@"¥%.2f",self.showMoney.doubleValue];
+    NSString *contentStr = [NSString stringWithFormat:@"共%ld件商品,合计:¥%.2f",self.products.count,self.showMoney.doubleValue];
     NSMutableAttributedString *attr = [NSString lh_attriStrWithprotocolStr:protocolStr content:contentStr protocolStrColor:APP_NAV_COLOR contentColor:kDarkGrayColor commonFont:FONT(14)];
     
-    if (self.derateMoney.floatValue>0) {
+    if (self.derateMoney.doubleValue>0) {
 
         if (self.isCanUseIntegral.integerValue == 1) {
             self.addtion_arr = @[@"",@"商家运费",@"减免活动",self.integralDisplayName];
 
-            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.floatValue],[NSString stringWithFormat:@"-¥%.2f",self.derateMoney.floatValue],@""];
+            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.doubleValue],[NSString stringWithFormat:@"-¥%.2f",self.derateMoney.doubleValue],@""];
         }else{
             self.addtion_arr = @[@"",@"商家运费",@"减免活动"];
-            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.floatValue],[NSString stringWithFormat:@"-¥%.2f",self.derateMoney.floatValue]];
+            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.doubleValue],[NSString stringWithFormat:@"-¥%.2f",self.derateMoney.doubleValue]];
         }
     }else{
         if (self.isCanUseIntegral.integerValue == 1) {
             self.addtion_arr = @[@"",@"商家运费",self.integralDisplayName];
-            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.floatValue],@""];
+            self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.doubleValue],@""];
         }else{
         self.addtion_arr = @[@"",@"商家运费"];
-        self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.floatValue]];
+        self.addtion_value_arr = @[attr,[NSString stringWithFormat:@"¥%.2f",self.freight.doubleValue]];
         }
     }
 }
